Fix issue when mouse move events are propagated in the document tree.
From lobovolution r49:

Fix issue when mouse move events are propagated in the document tree.
The RBlockViewport x,y position is always 0 even if the viewport
alignment is center or right which result in onMouseOver events to be
lost (ie. try adding a link with text-align: right and a:hover { color:
red; }: the event is not triggered). By always accepting the child
element (RBlockViewport) we can ensure that the event propagate further.