XHR: Now Working and without any permission exceptions.
Implemented Same Origin policy for XHR.

The call stack looks something like this when XHR.send() is called:
```
[Browser source ] XHR.send()
[JSEngine source] function interpreter
[Browser source ] callWithDomain()
[JSEngine source] runScript() or similar
[Browser source ] runScript()
```

A null code source also appears into the stack. It is not easy to trace
where the null domain comes from, but one of the causes seems to be the
logic in JVM for lambdas.

There were two main choices for implementing the permission model.

1. Let the callWithDomain() provide the protection. But it would require
the JSEngine scope and the null scope to be given all network
permissions.
2. Call doPrivileged() with restricted permissions in XHR.send()

I have taken the 2nd approach as it is doesn't require providing
permissions to JSEngine and null scope. This will be even more
significant when we add CORS support in the future, because first
approach would require giving a wildcard socket permission to JSEngine
and null scope.

In addition to the sandbox, I added a check for same origin policy in
XHR.send().