[security, cookies] added public-suffix-list library, fixed a hole in check
The library gives a more comprehensive list of public suffixes.

Also the check for suffix has an added condition:
the non-common part of suffix should be a '.' character.

This required a change to security policy. Now core permissions includes
the permission to read files on the classpath (not just directories).
Given that directories were readable, this is not a major change.
Secondly, tehse are only required for debugging in Eclipse; so could be
removed later.