for #149
In a previous commit, the hierarchy check was too stringent for
appendChild().

According to specs: If the given child is a reference to an existing
node in the document, appendChild() moves it from its current position
to the new position (i.e. there is no requirement to remove the node
from its parent node before appending it to some other node).

Also joined the following if with an else, since it was a mutually
exclusive condition